c_gpio.c: avoid use of fixed memory mapping
When a user-allocated memory pointer is passed to mmap() in the
Raspberry Pi, mmap() fails unless the pointer is a multiple of 16
kB (even though the page size in the system is 4 kB). The manual
page for mmap() says that requiring a fixed address for memory
mapping is less portable, so this option is discouraged in favor
of letting mmap() allocate the required memory itself.
In setup() function, avoid passing a fixed address to mmap(), and
rely on mmap() to allocate the memory instead.