c_gpio.c: fix check of mmap() return value
in setup() function, to check against failure of memory mapping
the return value of mmap() is cast to uint32_t and the cast value
is then checked for negative values, which can never happen
because an unsigned value cannot be negative.
Fix this by comparing the return value of mmap() with the
MAP_FAILED constant instead.